backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
pydantic==2.6.4
pydantic_core==2.16.3
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0
tsfresh==0.20.2
pandas==2.2.2
numpy==1.26.4
setuptools==69.5.1
